Initial pass at a conditional-store-based SFI for stores.  Changes: - Introduced a new pseudoinstruction, sfi_cstore_preamble (ARMInstrInfo.td).  Pseudoinstructions are a lame way to do this, but as in my last change, it's easier to let the assembler handle alignment than do it in LLVM for now. - Provided a definition for the sfi_cstore_preamble macro (ARMAsmPrinter.cpp). - Updated the ARMSFIPlacement pass to use the conditional store sandbox on instructions that are not currently predicate, but can be.  This actually leaves some cycles on the table by not applying it to EQ-predicated stores, but I need to finish liveness analysis for CPSR first.
As alluded to in that last item, there's a bug here.  We don't check liveness of CPSR before using the TST/STR sequence.  There's thus a chance that this transform violates correctness, if used on a program that
1. Produces a flag value in CPSR,
2. Does one or more non-predicated things, including a store,
3. Expects the CPSR flags to be intact for a predicated operation.

In practice, LLVM very rarely schedules instructions along a CPSR def-use chain, but it's still something I need to check.