Address the last bit of relocation flag related divergence betweeen LLVM and binutils.
With this patch, there are no functional differences between the .o
produced directly from LLVM versus the .s to .o via GNU as, for relocation tags
at least, for both PIC and non-PIC modes.

However, for PIC mode, I added a new switch on the ELFObjectwriter
-arm-elf-force-pic which forces the objectwriter to pretend that all relocs
are for PIC mode. This is due to an incompatibility of how relocs are emitted in
the .rodata section. PIC likes to emit relocs as section relative offsets, where as
non-pic seems to be fine with

The only detected differences in the .o machine code are in the values encoded
for movt instructions, but it is most likely a pair of related bugs in GNU
as/ld for ARM that is making the code work on the .s to .o path but is subtly
incompatible with the LLVM generated .o files

The only other differences in the .o files have to do with NOPs.  The current
NOP used in LLVM is a mov r0, r0 but GNU as like to use the real nop
instruction. This issue will be addressed separately.

So far the only failing test on PIC mode is env_var_test
On non-pic everything passes.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1588
BUG=http://code.google.com/p/nativeclient/issues/detail?id=424
TEST=utman.sh test-arm
TEST=utman.sh test-arm-pic