Don't inline __aeabi_read_tp
This simply reverts the parts of rev 241 and rev 242 that inline the
__aeabi_read_tp helper function rather than calling out to it.

The definition we were inlining was incorrect for the proper ABI, doing
a bogus subtraction of 16.  Removing the inlining for now lets us change
the definition in the library code along with cleaning up the rest of
the library code's TLS implementation.

We also want eventually to support the -mtls-use-call option to control
whether or not we do this inlining at all.  Without the inlining, it
becomes possible for the IRT to redefine __aeabi_read_tp to point to its
private "second" TLS area, which is necessary for correct operation of
the IRT without clobbering the application's TLS state.

After the TLS revamp settles, we can reintroduce the inlining, using the
correct definition (masking only, no subtraction), and conditionalizing
it on the new -mtls-use-call command-line option.